// src/types/enums/Errors.ts
export enum ErrorResponseNames {
  ValidationError = 'ValidationError',
  UsernameExistsException = 'UsernameExistsException',
  NotAuthorizedException = 'NotAuthorizedException',
  Unauthorized = 'Unauthorized',
  TokenExpiredError = 'TokenExpiredError',
  JsonWebTokenError = 'JsonWebTokenError',
  SessionExpired = 'SessionExpired',
  UnprocessableEntity = 'UnprocessableEntity',
  NotFound = 'NotFound',
  InternalError = 'InternalError',
}